+++ react/components/admin/redirects/UploadModal/UploadPrompt/index.tsx
@@ -7,7 +7,7 @@ import { Alert, AlertProps, Button, IconDelete, Radio } from 'vtex.styleguide'
 
 import UploadTableIcon from '../../../../icons/UploadTable'
 import { UploadActionType } from '../../mutations/SaveRedirectFromFile'
-import { CSV_HEADER, CSV_SEPARATOR } from '../../consts'
+import { getCSVHeader, CSV_SEPARATOR } from '../../consts'
 import { isRedirectType, validateRedirect } from './validateRedirect'
 import styles from './UploadPrompt.css'
 
@@ -17,10 +17,12 @@ interface Props {
   onDownloadTemplate: () => void
   saveRedirectFromFile: (
     parsedJsonFromCsv: Redirect[],
-    fileName: string
+    fileName: string,
+    hasMultipleBindings: boolean
   ) => void
   setUploadActionType: React.Dispatch<React.SetStateAction<UploadActionType>>
   uploadActionType: UploadActionType
+  hasMultipleBindings: boolean
 }
 
 const messages = defineMessages({
@@ -76,6 +78,16 @@ const messages = defineMessages({
     id:
       'admin/pages.admin.redirects.upload-modal.prompt.alert.validation-errors',
   },
+  missingBindingData: {
+    defaultMessage: 'Binding information missing from redirect {route}',
+    id:
+      'admin/pages.admin.redirects.upload-modal.prompt.alert.missing-binding-data',
+  },
+  extraBindingData: {
+    defaultMessage: 'Uneccessary binding information from redirect {route}',
+    id:
+      'admin/pages.admin.redirects.upload-modal.prompt.alert.extra-binding-data',
+  },
 })
 
 interface AlertState {
@@ -90,6 +102,7 @@ const UploadPrompt: React.FC<Props> = ({
   saveRedirectFromFile,
   setUploadActionType,
   uploadActionType,
+  hasMultipleBindings,
 }) => {
   const intl = useIntl()
   const [file, setFile] = useState<File | null>(null)
@@ -125,9 +138,9 @@ const UploadPrompt: React.FC<Props> = ({
 
       const reader = new FileReader()
       reader.onload = async () => {
-        const redirectKeys = CSV_HEADER.split(CSV_SEPARATOR) as Array<
-          keyof Redirect
-        >
+        const redirectKeys = getCSVHeader(hasMultipleBindings).split(
+          CSV_SEPARATOR
+        ) as Array<keyof Redirect>
         let parsedJsonFromCsv = null
         if (typeof reader.result === 'string') {
           parsedJsonFromCsv = reader.result
@@ -139,6 +152,24 @@ const UploadPrompt: React.FC<Props> = ({
               }
               const values = line.split(CSV_SEPARATOR)
               const redirect = zipObj(redirectKeys, values)
+              const wrongNumberOfPropeties =
+                redirectKeys.length != values.length
+              if (hasMultipleBindings && wrongNumberOfPropeties) {
+                nextValidationErrors = (nextValidationErrors || []).concat(
+                  intl.formatMessage(messages.missingBindingData, {
+                    route: redirect.from,
+                  })
+                )
+                return acc
+              }
+              if (!hasMultipleBindings && wrongNumberOfPropeties) {
+                nextValidationErrors = (nextValidationErrors || []).concat(
+                  intl.formatMessage(messages.extraBindingData, {
+                    route: redirect.from,
+                  })
+                )
+                return acc
+              }
               if (isRedirectType(redirect)) {
                 const validationError = validateRedirect(
                   redirect,
@@ -179,7 +210,7 @@ const UploadPrompt: React.FC<Props> = ({
       reader.readAsText(file)
       setFile(file)
     },
-    [intl, resetErrors, validationErrors]
+    [intl, resetErrors, validationErrors, hasMultipleBindings]
   )
 
   const { getRootProps, getInputProps } = useDropzone({
@@ -303,7 +334,11 @@ const UploadPrompt: React.FC<Props> = ({
           onClick={() =>
             parsedJsonFromCsv &&
             file &&
-            saveRedirectFromFile(parsedJsonFromCsv, file.name)
+            saveRedirectFromFile(
+              parsedJsonFromCsv,
+              file.name,
+              hasMultipleBindings
+            )
           }
         >
           {intl.formatMessage(messages.buttonImport)}