+++ react/components/admin/redirects/bulkUploadRedirects.ts
@@ -6,7 +6,7 @@ const NUMBER_OF_RETRIES = 3
 
 interface BulkUploadRedirectsArgs {
   data: Redirect[]
-  mutation: (data: Redirect[] | string[]) => Promise<void>
+  mutation: (data: Redirect[]) => Promise<void>
   isSave: boolean
   shouldUploadRef: MutableRefObject<boolean>
   updateProgress?: (processed: number) => void
@@ -15,7 +15,6 @@ interface BulkUploadRedirectsArgs {
 export default async function bulkUploadRedirects({
   data,
   mutation,
-  isSave,
   shouldUploadRef,
   updateProgress = () => {
     return undefined
@@ -25,12 +24,10 @@ export default async function bulkUploadRedirects({
 
   const redirectBatches = splitEvery(MAX_REDIRECTS_PER_REQUEST, data)
 
-  for (const data of redirectBatches) {
+  for (const payload of redirectBatches) {
     if (!shouldUploadRef.current) {
       break
     }
-
-    const payload = isSave ? data : data.map(({ from }) => from)
 
     for (let i = 1; i <= NUMBER_OF_RETRIES; i++) {
       try {